# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Customer 360 Intelligence Platform",
    "description": "\nCustomer 360 Intelligence Platform API.\n\nDeterministic customer data, agent-generated insights and grounded natural-language Q&A over a\nread-only banking dataset. Monetary values are integer cents internally and decimal strings at this\nboundary; every response carries a correlation ID and trace ID in `meta`.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/knowledge/search": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Hybrid knowledge retrieval with citations",
        "description": "Retrieve passages for ``q``, entitlement-filtered, and audit the documents returned.\n\nReranking is not enabled on this endpoint: it is the search surface, held to the dashboard-agent\nbudget. The Q&A path (Phase 9) enables reranking on its own route.",
        "operationId": "knowledge_search_knowledge_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Retrieval query (no PII).",
              "title": "Q"
            },
            "description": "Retrieval query (no PII)."
          },
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeDomain"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Restrict to these domains.",
              "title": "Domain"
            },
            "description": "Restrict to these domains."
          },
          {
            "name": "product_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Narrow to one product code.",
              "title": "Product Code"
            },
            "description": "Narrow to one product code."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_KnowledgeSearchPayload_"
                }
              }
            }
          },
          "503": {
            "description": "Audit unavailable (fail-closed)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge/documents/{doc_id}": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Document metadata",
        "description": "Return document metadata, or 404 if unknown or above the role's entitlement (17.8).\n\nAbsent and forbidden are collapsed into 404, so the endpoint is not an existence oracle for\nrestricted documents \u2014 the same discipline the customer endpoints apply.\n\nDocument metadata is institutional reference data (task 16.2): it changes only on re-ingestion,\nnot per request or per role. The response therefore carries a strong ``ETag`` and a short\n``Cache-Control`` so a repeat read can be a conditional ``304`` rather than a full round trip.",
        "operationId": "get_document_knowledge_documents__doc_id__get",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A specific version, or latest.",
              "title": "Version"
            },
            "description": "A specific version, or latest."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_DocumentPayload_"
                }
              }
            }
          },
          "304": {
            "description": "Not modified (client's ETag matches)"
          },
          "404": {
            "description": "Unknown document or not entitled"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentPayload": {
        "properties": {
          "doc_id": {
            "type": "string",
            "title": "Doc Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "domain": {
            "$ref": "#/components/schemas/KnowledgeDomain"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "effective_from": {
            "type": "string",
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction"
          },
          "access_level": {
            "type": "string",
            "title": "Access Level"
          },
          "product_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Code"
          },
          "business_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Group"
          }
        },
        "type": "object",
        "required": [
          "doc_id",
          "title",
          "domain",
          "version",
          "effective_from",
          "access_level"
        ],
        "title": "DocumentPayload",
        "description": "``GET /knowledge/documents/{doc_id}`` payload: document metadata (design \u00a76.3)."
      },
...
      "Envelope_DocumentPayload_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DocumentPayload"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[DocumentPayload]"
      },
...
      "Envelope_KnowledgeSearchPayload_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/KnowledgeSearchPayload"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "Envelope[KnowledgeSearchPayload]"
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "UNAUTHENTICATED",
          "ENTITLEMENT_DENIED",
          "CUSTOMER_NOT_FOUND",
          "VALIDATION_ERROR",
          "AGENT_TIMEOUT",
          "UPSTREAM_UNAVAILABLE",
          "RATE_LIMITED",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "description": "The complete error code set from design \u00a76.3."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KnowledgeDomain": {
        "type": "string",
        "enum": [
          "product_catalog",
          "policy",
          "procedure",
          "offer_terms",
          "playbook",
          "compliance"
        ],
        "title": "KnowledgeDomain",
        "description": "The six knowledge domains (design \u00a79.2). Mirrors the ``kb_document.domain`` CHECK set."
      },
      "KnowledgePassage": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "doc_id": {
            "type": "string",
            "title": "Doc Id"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "section_path": {
            "type": "string",
            "title": "Section Path"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "effective_from": {
            "type": "string",
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          }
        },
        "type": "object",
        "required": [
          "text",
          "score",
          "doc_id",
          "version",
          "section_path",
          "title",
          "effective_from",
          "chunk_id"
        ],
        "title": "KnowledgePassage",
        "description": "One cited passage from the tool: text plus everything needed to cite it (17.5, 17.6)."
      },
      "KnowledgeSearchPayload": {
        "properties": {
          "passages": {
            "items": {
              "$ref": "#/components/schemas/KnowledgePassage"
            },
            "type": "array",
            "title": "Passages",
            "default": []
          },
          "candidate_count": {
            "type": "integer",
            "title": "Candidate Count",
            "default": 0
          },
          "reranked": {
            "type": "boolean",
            "title": "Reranked",
            "default": false
          },
          "no_guidance_found": {
            "type": "boolean",
            "title": "No Guidance Found",
            "default": false
          }
        },
        "type": "object",
        "title": "KnowledgeSearchPayload",
        "description": "``GET /knowledge/search`` payload: cited passages and retrieval counters (17.5, 17.6)."
      },
...
      "Meta": {
        "properties": {
          "correlation_id": {
            "type": "string",
            "title": "Correlation Id"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          },
          "as_of": {
            "type": "string",
            "format": "date-time",
            "title": "As Of"
          },
          "masked_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Masked Fields"
          },
          "computed_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Computed Fields"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ModuleError"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "correlation_id",
          "as_of"
        ],
        "title": "Meta",
        "description": "Envelope metadata."
      },
      "ModuleError": {
        "properties": {
          "module": {
            "type": "string",
            "title": "Module"
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "module",
          "code",
          "message"
        ],
        "title": "ModuleError",
        "description": "A single domain module's failure, surfaced without failing the whole response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}